fix: circular dependency and repeated code
To fix the circular dependency, I added the comparison by the
`.currency` attribute, instead of comparing by class.
This was an excelent idea that blowed my mind (that shouldn't happened
but it was unexpected for me when I read that).

As there was a comping problem (red) I didn't implement tests for
currency... I'll implement this in the next commit. But not sure this is
fine.